decoupled ECS from scene
