+++ src/Relatus/ECS/Systems/SimpleShapeSystem.cs
@@ -33,16 +33,16 @@ namespace Relatus.ECS
         /// By default, this system will handle any entity that includes the following components: <see cref="CPosition"/>, <see cref="CDimension"/>, <see cref="CTransform"/>, <see cref="CColor"/>,
         /// and a custom <see cref="IComponent"/> that acts as a tag specifically for this system.
         /// </summary>
-        /// <param name="scene">The scene this system will exist in.</param>
+        /// <param name="factory">The scene this system will exist in.</param>
         /// <param name="geometry">The shape data that this system will focus on and draw.</param>
         /// <param name="shapeTag">The type of the custom <see cref="IComponent"/> that acts as a tag specifically for this system.</param>
         /// <param name="tasks">The total amount of tasks to divide the update cycle into. Assigning more than one task allows entities to be updated asynchronously.</param>
-        public SimpleShapeSystem(Scene scene, GeometryData geometry, Type shapeTag, uint tasks) : base(scene, tasks)
+        public SimpleShapeSystem(MorroFactory factory, GeometryData geometry, Type shapeTag, uint tasks) : base(factory, tasks)
         {
             Require(typeof(CPosition), typeof(CDimension), typeof(CTransform), typeof(CColor), shapeTag);
 
             this.geometry = geometry;
-            vertexBuffer = new VertexTransformColor[scene.EntityCapacity];
+            vertexBuffer = new VertexTransformColor[factory.EntityCapacity];
         }
 
         public override void UpdateEntity(int entity)
@@ -87,10 +87,10 @@ namespace Relatus.ECS
 
         public override void Update()
         {
-            positions = scene.GetData<CPosition>();
-            dimensions = scene.GetData<CDimension>();
-            transforms = scene.GetData<CTransform>();
-            colors = scene.GetData<CColor>();
+            positions = factory.GetData<CPosition>();
+            dimensions = factory.GetData<CDimension>();
+            transforms = factory.GetData<CTransform>();
+            colors = factory.GetData<CColor>();
 
             Array.Clear(vertexBuffer, 0, vertexBuffer.Length);
